+++ optax/_src/alias.py
@@ -722,106 +722,6 @@ def adamw(
   )
 
 
-def adopt(
-    learning_rate: base.ScalarOrSchedule,
-    b1: float = 0.9,
-    b2: float = 0.9999,
-    eps: float = 1e-6,
-    mu_dtype: Optional[Any] = None,
-    *,
-    nesterov: bool = False,
-    use_clipping: bool = True,
-) -> base.GradientTransformationExtraArgs:
-  r"""ADOPT: Modified Adam optimizer that can converge with any β2 value.
-
-  ADOPT (Adaptive Optimization with Provable Theoretical guarantees) is a modified 
-  version of Adam that ensures convergence with any β2 value while maintaining the 
-  optimal convergence rate. This implementation includes an optional clipping 
-  operation to improve stability, especially in early training stages.
-
-  The key difference from Adam is that ADOPT modifies the update rule to avoid 
-  potential instability issues, particularly when some gradient elements are near 
-  zero at initialization. This can happen when parameters (e.g., the last layer of 
-  a neural network) are initialized with zeros.
-
-  Let :math:`\alpha_t` represent the learning rate and :math:`\beta_1, \beta_2`,
-  :math:`\varepsilon`, :math:`\bar{\varepsilon}` represent the arguments
-  ``b1``, ``b2``, ``eps`` and ``eps_root`` respectively. The learning rate is
-  indexed by :math:`t` since the learning rate may also be provided by a
-  schedule function.
-
-  The ``init`` function of this optimizer initializes an internal state
-  :math:`S_0 := (m_0, v_0) = (0, 0)`, representing initial estimates for the
-  first and second moments. At step :math:`t`, the ``update`` function computes:
-
-  .. math::
-
-    \begin{align*}
-      m_t &\leftarrow \beta_1 \cdot m_{t-1} + (1-\beta_1) \cdot g_t \\
-      v_t &\leftarrow \beta_2 \cdot v_{t-1} + (1-\beta_2) \cdot {g_t}^2 \\
-      \hat{m}_t &\leftarrow m_t / {(1-\beta_1^t)} \\
-      \hat{v}_t &\leftarrow v_t / {(1-\beta_2^t)} \\
-    \end{align*}
-
-  With clipping enabled (default), ADOPT applies a clipping operation to improve 
-  stability, particularly in early training stages.
-
-  Args:
-    learning_rate: A global scaling factor, either fixed or evolving along
-      iterations with a scheduler.
-    b1: Exponential decay rate to track the first moment of past gradients.
-    b2: Exponential decay rate to track the second moment of past gradients.
-      Unlike Adam, ADOPT can converge with any value of b2.
-    eps: A small constant applied to denominator outside of the square root
-      to avoid dividing by zero when rescaling.
-    eps_root: A small constant applied to denominator inside the square root,
-      to avoid dividing by zero when rescaling.
-    mu_dtype: Optional `dtype` to be used for the first order accumulator; if
-      `None` then the `dtype` is inferred from `params` and `updates`.
-    nesterov: Whether to use Nesterov momentum.
-    use_clipping: Whether to apply clipping to improve stability. Recommended
-      to keep as True, especially for training from scratch.
-
-  Returns:
-    The corresponding :class:`optax.GradientTransformationExtraArgs`.
-
-  Examples:
-    >>> import optax
-    >>> import jax
-    >>> import jax.numpy as jnp
-    >>> def f(x): return jnp.sum(x ** 2)  # simple quadratic function
-    >>> solver = optax.adopt(learning_rate=0.003)
-    >>> params = jnp.array([1., 2., 3.])
-    >>> print('Objective function: ', f(params))
-    Objective function:  14.0
-    >>> opt_state = solver.init(params)
-    >>> for _ in range(5):
-    ...  grad = jax.grad(f)(params)
-    ...  updates, opt_state = solver.update(grad, opt_state, params)
-    ...  params = optax.apply_updates(params, updates)
-    ...  print('Objective function: {:.2E}'.format(f(params)))
-    Objective function: 1.40E+01
-    Objective function: 1.39E+01
-    Objective function: 1.39E+01
-    Objective function: 1.39E+01
-    Objective function: 1.38E+01
-
-  References:
-    Taniguchi et al, `ADOPT: Modified Adam Can Converge with Any β2 with the 
-    Optimal Rate <https://arxiv.org/abs/2403.00855>`_, NeurIPS 2024
-  """
-  return combine.chain(
-      transform.scale_by_adopt(
-          b1=b1,
-          b2=b2,
-          eps=eps,
-          mu_dtype=mu_dtype,
-          nesterov=nesterov,
-          use_clipping=use_clipping,
-      ),
-      transform.scale_by_learning_rate(learning_rate),
-  )
-
 
 nadamw = functools.partial(adamw, nesterov=True)
 nadamw.__doc__ = (
